decimal places at jitter and ping
better values with --share
The browser version doesn't really seem to use decimal places at ping but this
limits/extended ping and jitter to .2 decimal places at --csv / --json and stdout.
this prevent 1.0909090909090908 ping times within csv and json and extend those values in rest of the data.
The ping value could be adjusted to .0 to match the web version but this way the values are more accurate. +1 for the cli.